Fix for the bug when deleting DNC numbers

+++ goLists/goActionDNC.php
@@ -79,8 +79,10 @@
 				}
 			} else {
 				if ($stage == "DELETE"){
-					$query = mysqli_query($link, "DELETE FROM vicidial_campaign_dnc WHERE phone_number='$dnc' AND campaign_id='".$campaign_id."'");
-					log_action ('DELETE', $user_id, $ip_address, date('Y-m-d H:i:s'), "Deleted DNC Number $dnc from Campaign ".$campaign_id , $usergroup, "DELETE FROM vicidial_campaign_dnc WHERE phone_number='$dnc' AND campaign_id='".$campaign_id."';");
+					if ($cdnc_exist > 0) {
+						$query = mysqli_query($link, "DELETE FROM vicidial_campaign_dnc WHERE phone_number='$dnc' AND campaign_id='".$campaign_id."'");
+						log_action ('DELETE', $user_id, $ip_address, date('Y-m-d H:i:s'), "Deleted DNC Number $dnc from Campaign ".$campaign_id , $usergroup, "DELETE FROM vicidial_campaign_dnc WHERE phone_number='$dnc' AND campaign_id='".$campaign_id."';");
+					}
 					
 					if ($campaign_id === '' && $idnc_exist > 0) {
 						$query = mysqli_query($link, "DELETE FROM vicidial_dnc WHERE phone_number='$dnc';");
